Support serverless deployments with `@search` pragma
We're going to add some more configs to the `@search` pragma to
adjust sandbox settings, so we can no longer use the presence or
absence of the pragma itself to determine whether to deploy to
service mode or serverless mode.

Instead, check whether we are in service or serverless mode by
looking for one of the config keys that is required for service
mode.